requirements.txt
playwright==1.44.0
pillow
requests
beautifulsoup4
